Updated calculations in the nav calcs and Madgwick filter drivers. Added a getter for poition uncertainty in the M8Q driver.

+++ tools/nav_calcs.h
@@ -284,9 +284,9 @@ public:
      * 
      * @see TrueNorthHeading 
      * 
-     * @param tn_offset : offset between magnetic and true north (degrees*10) 
+     * @param tn_offset : offset between magnetic and true north (degrees) 
      */
-    void SetTnOffset(int16_t tn_offset);
+    void SetTnOffset(float tn_offset);
 
     /**
      * @brief Set the Kalman Pose Data objectSet Kalman filter data
